+++ CHANGES.rst
@@ -1,4 +1,10 @@
+0.0.2 (02/23/2011)
+==================
+* using SmartDisplay
+* new: bgcolor option
+* new: timeout option
+
 0.0.1 (02/12/2011)
 ==================
 
-- Initial release
+* Initial release